Switched to composite keys for locations instead of sequential ids for easier syncing with backend in future

+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/database/dao/DailyForecastDAO.java
@@ -46,7 +46,7 @@ public class DailyForecastDAO extends ForecastDAO<DailyForecast> {
    * @return A list of all DailyForecast objects in the database that are associated with the specified location.
    */
   @Deprecated
-  public List<DailyForecast> getByLocationId(int location_id) {
+  public List<DailyForecast> getByLocationId(long location_id) {
     return new DailyForecastQuery()
         .whereLocationId(location_id)
         .getResults();
